requirements.txt
allure-pytest==2.13.2
Faker==22.6.0
pytest==8.0.0
requests==2.31.0
selenium==4.17.2